+++ frontend/src/employee-frontend/api/child/placements.ts
@@ -4,10 +4,11 @@
 
 import { Failure, Result, Success } from 'lib-common/api'
 import { client } from '../../api/client'
-import { Placement, PlacementType } from '../../types/child'
+import { Placement } from '../../types/child'
 import { UUID } from '../../types'
 import { JsonOf } from 'lib-common/json'
 import LocalDate from 'lib-common/local-date'
+import { PlacementType } from 'lib-common/api-types/serviceNeed/common'
 
 export interface PlacementCreate {
   childId: UUID
