+++ frontend/src/employee-frontend/api/daycare.ts
@@ -6,13 +6,13 @@ import { client } from './client'
 import { Failure, Result, Success } from 'lib-common/api'
 import { Unit } from '../types/invoicing'
 import { CareArea } from '../types/unit'
-import { PlacementType } from '../types/child'
 import { JsonOf } from 'lib-common/json'
 import LocalDate from 'lib-common/local-date'
 import {
   deserializePublicUnit,
   PublicUnit
 } from 'lib-common/api-types/units/PublicUnit'
+import { PlacementType } from 'lib-common/api-types/serviceNeed/common'
 
 export async function getUnits(
   areas: string[],
