+++ frontend/src/lib-common/api-types/serviceNeed/common.ts
@@ -1,6 +1,4 @@
 import { UUID } from 'lib-common/types'
-import { PlacementType } from '../application/enums'
-
 export interface ServiceNeedOptionPublicInfo {
   id: UUID
   name: string
@@ -11,3 +9,16 @@ export interface ServiceNeedOptionSummary {
   id: UUID
   name: string
 }
+
+export type PlacementType =
+  | 'CLUB'
+  | 'DAYCARE'
+  | 'DAYCARE_PART_TIME'
+  | 'DAYCARE_FIVE_YEAR_OLDS'
+  | 'DAYCARE_PART_TIME_FIVE_YEAR_OLDS'
+  | 'PRESCHOOL'
+  | 'PRESCHOOL_DAYCARE'
+  | 'PREPARATORY'
+  | 'PREPARATORY_DAYCARE'
+  | 'TEMPORARY_DAYCARE'
+  | 'TEMPORARY_DAYCARE_PART_DAY'
